rename pack/chips/sun50iw6p1/configs/petrel-p1 to petrel-p1-orig
petrel-p1 is just one of the boards using sun50iw6p1. Ideally we
should be able to switch by selecting which directory to use.

But 'pack' uses Android build variable PRODUCT_BOARD to determine
the board. So unless we create a separate product under Android
we are effectively hardcoded to petrel-p1 as that is the name of
the product under Android build.
(see device/softwinner/common/vendorsetup.sh)

As a hack/expedience, we we make 'petrel-p1' as a symbolic link to
the actual directory.